+++ desktop/cmp/dash/canvas/impl/DashCanvasView.ts
@@ -34,13 +34,14 @@ export const dashCanvasView = hoistCmp.factory({
     model: uses(DashCanvasViewModel, {publishMode: 'limited'}),
 
     render({model, className}) {
-        const {viewSpec, ref, hidePanelHeader, headerItems, autoHeight} = model,
+        const {viewSpec, ref, hidePanelHeader, headerItems, autoHeight, titleDetails, title, icon} =
+                model,
             headerProps = hidePanelHeader
                 ? {}
                 : {
                       compactHeader: true,
-                      title: model.title,
-                      icon: model.icon,
+                      title: title + (titleDetails ? ' ' + titleDetails : ''),
+                      icon,
                       headerItems: [...headerItems, headerMenu({model})]
                   };
         return panel({